meta: factor out SolverScope around ix and log_file

+++ meta.py
@@ -381,6 +381,43 @@ class SolverLibrary:
         yield ui, solver
         ui.interact(solver)
 
+@final
+class SolverScope:
+    def __init__(self,
+                 lib: SolverLibrary,
+                 ix: Iterable[int]|None=None):
+        self.lib = lib
+        self.ix = list(range(len(lib)) if ix is None else ix)
+        self.log_file = [lib.proto[i].log_file for i in self.ix]
+
+    def __len__(self):
+        return len(self.ix)
+
+    def __iter__(self) -> Generator[int]:
+        yield from self.ix
+
+    def __contains__(self, i: int):
+        return i in self.ix
+
+    def lookup(self, solver_i: int|None=None, name: str=''):
+        if solver_i is None:
+            solver_i = self.lib.by_name.get(name)
+        if solver_i is not None:
+            for j, i in enumerate(self.ix):
+                if i == solver_i:
+                    return j
+
+    def run(self,
+            ui: PromptUI,
+            name: str='',
+            solver_i: int|None=None,
+            log_file: str|None=None):
+        if log_file is None:
+            j = self.lookup(solver_i, name)
+            if j is not None:
+                log_file = self.log_file[j]
+        return self.lib.run(ui, solver_i=solver_i, name=name, log_file=log_file)
+
 solvers = SolverLibrary()
 
 from binartic import Search as Binartic
@@ -528,9 +565,9 @@ class Meta(Arguable):
 
     def __init__(self):
         super().__init__()
-        self.solver_log = [proto.log_file for proto in solvers.proto]
         self.report = Report()
         self.prompt.mess = self.prompt_mess
+        self.solvers = SolverScope(solvers)
 
         root = self.prompt
 
@@ -700,7 +737,7 @@ class Meta(Arguable):
             if day == today:
                 solved = all(
                     solver_i in solves
-                    for solver_i in range(len(solvers)))
+                    for solver_i in self.solvers)
                 if solved: # TODO and shared
                     prune = True
                 ui.write(f' today solved: {solved}')
@@ -866,15 +903,14 @@ class Meta(Arguable):
         '''
         run solver
         '''
-        with solvers.run(ui, solver_i=solver_i):
+        with self.solvers.run(ui, solver_i=solver_i):
             pass
 
     def do_sol_cont(self, solver_i: int, ui: PromptUI):
         '''
         continue solver run
         '''
-        log_file = self.solver_log[solver_i]
-        with solvers.run(ui, solver_i=solver_i, log_file=log_file):
+        with self.solvers.run(ui, solver_i=solver_i):
             pass
 
     def do_sol_edit(self, solver_i: int, ui: PromptUI):
@@ -882,65 +918,77 @@ class Meta(Arguable):
         open solver log in $EDITOR
         '''
         editor = os.environ.get('EDITOR', 'vi')
-        log_file = self.solver_log[solver_i]
-        with ui.check_proc(subprocess.Popen((editor, log_file))):
-            pass
+        j = self.solvers.lookup(solver_i)
+        if j is not None:
+            log_file = self.solvers.log_file[j]
+            with ui.check_proc(subprocess.Popen((editor, log_file))):
+                pass
 
     def do_sol_cur(self, solver_i: int, ui: PromptUI):
         '''
         use current log file
         '''
-        proto = solvers.proto[solver_i]
-        prior = self.solver_log[solver_i]
-        log_file = proto.log_file
-        self.solver_log[solver_i] = log_file
-        action = 'Reset' if prior != log_file else 'Current'
-        ui.print(f'{action} log file: {log_file}')
+        j = self.solvers.lookup(solver_i)
+        if j is not None:
+            proto = solvers.proto[solver_i]
+            prior = self.solvers.log_file[j]
+            log_file = proto.log_file
+            self.solvers.log_file[j] = log_file
+            action = 'Reset' if prior != log_file else 'Current'
+            ui.print(f'{action} log file: {log_file}')
 
     def do_sol_last(self, solver_i: int, ui: PromptUI):
         '''
         use latest stored solver log
         '''
         proto = solvers.proto[solver_i]
-        log_file = proto.find_prior_log(ui, puzzle_id=None)
-        if log_file is None:
-            ui.print(f'! could not find last log file')
-            return
-        ui.print(f'Found last log_file: {log_file}')
-        self.solver_log[solver_i] = log_file
+        j = self.solvers.lookup(solver_i)
+        if j is not None:
+            log_file = proto.find_prior_log(ui, puzzle_id=None)
+            if log_file is None:
+                ui.print(f'! could not find last log file')
+                return
+            ui.print(f'Found last log_file: {log_file}')
+            self.solvers.log_file[j] = log_file
 
     def do_sol_ls(self, solver_i: int, ui: PromptUI):
         '''
         list and select from stored solver logs
         '''
         proto = solvers.proto[solver_i]
-        log_file = proto.find_prior_log(ui, puzzle_id='*')
-        if log_file is not None:
-            ui.print(f'Selected log_file: {log_file}')
-            self.solver_log[solver_i] = log_file
+        j = self.solvers.lookup(solver_i)
+        if j is not None:
+            log_file = proto.find_prior_log(ui, puzzle_id='*')
+            if log_file is not None:
+                ui.print(f'Selected log_file: {log_file}')
+                self.solvers.log_file[j] = log_file
 
     def do_sol_rm(self, solver_i: int, ui: PromptUI):
         '''
         remove solver log file
         '''
-        log_file = self.solver_log[solver_i]
-        ui.print(f'+ rm {log_file}')
-        try:
-            os.unlink(log_file)
-        except OSError as err:
-            ui.print(f'! {err}')
+        j = self.solvers.lookup(solver_i)
+        if j is not None:
+            log_file = self.solvers.log_file[j]
+            ui.print(f'+ rm {log_file}')
+            try:
+                os.unlink(log_file)
+            except OSError as err:
+                ui.print(f'! {err}')
 
     def do_sol_tail(self, solver_i: int, ui: PromptUI):
         '''
         show last N lines from solver log file
         '''
-        log_file = self.solver_log[solver_i]
-        tail_n = (
-            3 if ui.screen_lines < 10 else
-            10 if ui.screen_lines < 20 else
-            ui.screen_lines//2)
-        with ui.check_proc(subprocess.Popen(('tail', f'-n{tail_n}', log_file))):
-            pass
+        j = self.solvers.lookup(solver_i)
+        if j is not None:
+            log_file = self.solvers.log_file[j]
+            tail_n = (
+                3 if ui.screen_lines < 10 else
+                10 if ui.screen_lines < 20 else
+                ui.screen_lines//2)
+            with ui.check_proc(subprocess.Popen(('tail', f'-n{tail_n}', log_file))):
+                pass
 
     def do_review(self, ui: PromptUI):
         try:
@@ -976,22 +1024,22 @@ class Meta(Arguable):
                     continue
                 if day is not None and head:
                     continue
-                proto = solvers.proto[solver_i]
+                proto = self.solvers.lib.proto[solver_i]
                 if proto.site_env != 'prod':
                     continue
                 if once(solver_i):
                     yield solver_i
 
         for solver_i in candidates():
-            name = solvers.name[solver_i]
-            proto = solvers.proto[solver_i]
+            name = self.solvers.lib.name[solver_i]
+            proto = self.solvers.lib.proto[solver_i]
             desc = (
                 f'[ {self.solvers.ix.index(solver_i)} / {len(self.solvers)} ] {name}'
                 if solver_i == solver_j else
                 f'{name} variant {proto.note_slug[0]}')
 
             ui.print(f'Running {desc}')
-            with solvers.run(ui, solver_i=solver_i):
+            with self.solvers.run(ui, solver_i=solver_i):
                 return
 
         ui.print('! all solvers reported, specify particular?')
@@ -1006,6 +1054,7 @@ class Meta(Arguable):
             ui.print(f'{solver_i + 1}. {name} site:{proto.site!r} slug:{note!r}')
 
     def read_status(self, ui: PromptUI, verbose: int=0):
+        solvers = self.solvers.lib
         solver_notes = tuple(proto.note_slug[0] for proto in solvers.proto)
         solver_heads = tuple(proto.header_slug[0] for proto in solvers.proto)
 
@@ -1057,13 +1106,12 @@ class Meta(Arguable):
                 if not note:
                     ui.print(f'Missing {slug!r}')
 
-        for solver_i in range(len(solver_name)):
+        for solver_i in self.solvers:
             dd_n = note_days[solver_i]
             day = days[dd_n-1] if dd_n else None
             if verbose > 1:
                 ui.print(f'* [{solver_i}] day=[{dd_n-1}]={day} base')
             yield solver_i, day, notes[solver_i], heads[solver_i], bodys[solver_i]
-
 
     def do_status(self, ui: PromptUI):
         '''
